Remove skin editor functionality from ToolsDashboard
Updated update-check logic to support offline mode (no internet)

Replaced the "L" by the actual logo of the app

Moved the skin editor to the skin tab
