+++ src/App.tsx
@@ -14,8 +14,7 @@ const ServerSettings = React.lazy(() => import('./pages/ServerSettings'));
 const ServerSearch = React.lazy(() => import('./pages/ServerSearch'));
 const ServerLibrary = React.lazy(() => import('./pages/ServerLibrary'));
 const InstanceDetails = React.lazy(() => import('./pages/InstanceDetails'));
-const Client = React.lazy(() => import('./pages/Client'));
-const ClientMods = React.lazy(() => import('./pages/ClientMods'));
+
 const ToolsDashboard = React.lazy(() => import('./pages/ToolsDashboard'));
 const Extensions = React.lazy(() => import('./pages/Extensions'));
 const Login = React.lazy(() => import('./pages/Login'));
@@ -1173,31 +1172,7 @@ function App() {
             }
         }
 
-        if (currentMode === 'client' && isFeatureEnabled('openClientPage')) {
-            if (currentView === 'open-client') {
-                return <Client />;
-            }
 
-            if (currentView === 'skins' && !isGuest) {
-                return <Skins onLogout={handleLogout} onProfileUpdate={setUserProfile} />;
-            }
-
-            if (currentView === 'extensions') {
-                return <Extensions />;
-            }
-
-            if (currentView === 'styling') {
-                return <Styling />;
-            }
-
-            if (currentView === 'mods') {
-                return <ClientMods />;
-            }
-
-            if (currentView === 'settings') {
-                return <Settings mode="client" onRestartGuide={() => handleRestartGuide('client')} onClose={handleCloseSettingsOverlay} disableClose={isGuideRunning} />;
-            }
-        }
 
         if (currentMode === 'tools') {
             if (currentView === 'tools-dashboard') {
